fixed GraceCL nit that stopped lisp from launching, fixe int arg in Midi.cpp
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1694 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/Midi.cpp
@@ -2461,7 +2461,7 @@ MidiFileInfoComponent::~MidiFileInfoComponent()
 
 void MidiFileInfoComponent::resized()
 {
-  float y=12;
+  int y=12;
   // only active is saving file
   if (savebutton)
     {
